+++ app/src/main/java/com/jingom/seizetheday/domain/model/AttachedImage.kt
@@ -1,7 +1,7 @@
 package com.jingom.seizetheday.domain.model
 
 data class AttachedImage(
-	val id: Int,
-	val thanksRecordId: Int,
+	val id: Long,
+	val thanksRecordId: Long,
 	val imageUri: String
 )
